refactored structure of packages
