+++ android/src/space/hypeo/mankomania/AndroidLauncher.java
@@ -8,9 +8,6 @@ import com.badlogic.gdx.backends.android.AndroidApplicationConfiguration;
 import java.util.LinkedList;
 import java.util.List;
 
-import space.hypeo.networking.network.IDeviceStateSubscriber;
-import space.hypeo.networking.network.IDeviceStatePublisher;
-
 public class AndroidLauncher extends AndroidApplication implements IDeviceStatePublisher {
 
     Mankomania mankomania = null;
