+++ core/src/space/hypeo/mankomania/IDeviceStatePublisher.java
@@ -1,4 +1,4 @@
-package space.hypeo.networking.network;
+package space.hypeo.mankomania;
 
 /**
  * Created by pichlermarc on 18.05.2018.
