+++ core/src/space/hypeo/mankomania/Mankomania.java
@@ -6,8 +6,6 @@ import com.badlogic.gdx.graphics.GL20;
 import com.badlogic.gdx.graphics.OrthographicCamera;
 import com.badlogic.gdx.utils.viewport.ExtendViewport;
 
-import space.hypeo.networking.network.IDeviceStatePublisher;
-
 
 /**
  * The GDX Game class, called from the android project.
