+++ core/src/space/hypeo/mankomania/player/IPlayerConnector.java
@@ -1,9 +1,11 @@
-package space.hypeo.networking.network;
+package space.hypeo.mankomania.player;
 
 
 /**
  * Created by pichlermarc on 07.04.2018.
  */
+
+import space.hypeo.mankomania.player.Lobby;
 
 /**
  * Interface provides service methods to fetch data that are relevant for each player
