+++ core/src/space/hypeo/mankomania/stages/MainMenuStage.java
@@ -14,12 +14,12 @@ import com.badlogic.gdx.scenes.scene2d.utils.TextureRegionDrawable;
 import com.badlogic.gdx.utils.Align;
 import com.badlogic.gdx.utils.viewport.Viewport;
 
-import space.hypeo.player.PlayerFactory;
-import space.hypeo.player.PlayerManager;
+import space.hypeo.mankomania.player.PlayerFactory;
+import space.hypeo.mankomania.player.PlayerManager;
 import space.hypeo.mankomania.StageFactory;
 import space.hypeo.mankomania.StageManager;
 import space.hypeo.mankomania.actors.common.RectangleActor;
-import space.hypeo.networking.network.IDeviceStatePublisher;
+import space.hypeo.mankomania.IDeviceStatePublisher;
 import space.hypeo.networking.network.Role;
 
 /**
