+++ core/src/space/hypeo/networking/packages/PlayerDisconnect.java
@@ -1,6 +1,6 @@
 package space.hypeo.networking.packages;
 
-import space.hypeo.networking.network.PlayerSkeleton;
+import space.hypeo.mankomania.player.PlayerSkeleton;
 
 /**
  * This class is only a wrapper class for player to send over the network and invoke right action.
